119 backend refactor all typings to use zod (#154)
* feat: create media and user zod type

* feat: create client additional info zod object

* feat: create semester schema

also refactor all types to use satisfies, also use nativeEnum for ease

* feat: create project zod schema

* feat: create semester project zod type

* feat: create form schemas

- Replaced manual QuestionResponse interface with Zod-based schema.
- Added FormQuestionSchema, QuestionResponseSchema, FormSchema, and
  FormResponseSchema to define related types.
- Updated imports and references to align with the new schema structure.

* refactor: remove code smells and move after each test code to overall after each hook

* fix: tests in form service to have an overall describe

* style: run prettier

* refactor: move restricted route to correct folder

* refactor: move endpoints to correct folder

* refactor: remove get method from projectId path

* refactor: move admin routes to admin folder and rename path for convention
