+++ src/app/api/admin/projects/[id]/route.ts
@@ -1,36 +1,10 @@
-import { NextRequest, NextResponse } from 'next/server'
 import { StatusCodes } from 'http-status-codes'
-import ProjectService from '@/data-layer/services/ProjectService'
+import { NextRequest, NextResponse } from 'next/server'
 import { NotFound } from 'payload'
-import { UpdateProjectRequestBody } from '@/types/request-models/ProjectRequests'
 import { ZodError } from 'zod'
 
-/**
- * Fetches a project by its ID.
- * @param req - The request object.
- * @param params - The parameters object containing the project ID.
- */
-
-export const GET = async (
-  req: NextRequest,
-  { params }: { params: Promise<{ projectId: string }> },
-): Promise<Response> => {
-  const { projectId } = await params
-  const projectService = new ProjectService()
-
-  try {
-    const data = await projectService.getProjectById(projectId)
-    return NextResponse.json({ data })
-  } catch (error) {
-    if (error instanceof NotFound) {
-      return NextResponse.json({ error: 'Project not found' }, { status: StatusCodes.NOT_FOUND })
-    }
-    return NextResponse.json(
-      { error: 'Internal Server Error' },
-      { status: StatusCodes.INTERNAL_SERVER_ERROR },
-    )
-  }
-}
+import ProjectService from '@/data-layer/services/ProjectService'
+import { UpdateProjectRequestBody } from '@/types/request-models/ProjectRequests'
 
 /**
  * Patches a project by its ID.
@@ -39,13 +13,13 @@ export const GET = async (
  */
 export const PATCH = async (
   req: NextRequest,
-  { params }: { params: Promise<{ projectId: string }> },
+  { params }: { params: Promise<{ id: string }> },
 ): Promise<NextResponse> => {
-  const { projectId } = await params
+  const { id } = await params
   const projectService = new ProjectService()
   try {
     const body = UpdateProjectRequestBody.parse(await req.json())
-    const data = await projectService.updateProject(projectId, body)
+    const data = await projectService.updateProject(id, body)
     return NextResponse.json({ data: data })
   } catch (error) {
     if (error instanceof NotFound) {
@@ -71,13 +45,13 @@ export const PATCH = async (
  */
 export const DELETE = async (
   req: NextRequest,
-  { params }: { params: Promise<{ projectId: string }> },
+  { params }: { params: Promise<{ id: string }> },
 ): Promise<NextResponse> => {
-  const { projectId } = await params
+  const { id } = await params
   const projectService = new ProjectService()
 
   try {
-    await projectService.deleteProject(projectId)
+    await projectService.deleteProject(id)
     return new NextResponse(null, { status: StatusCodes.NO_CONTENT })
   } catch (error) {
     if (error instanceof NotFound) {
