+++ src/app/api/admin/semesters/[id]/route.ts
@@ -1,8 +1,42 @@
 import { StatusCodes } from 'http-status-codes'
 import { NextRequest, NextResponse } from 'next/server'
 import { NotFound } from 'payload'
+import { ZodError } from 'zod'
 
 import SemesterService from '@/data-layer/services/SemesterService'
+import { UpdateSemesterRequestBody } from '@/types/request-models/SemesterRequests'
+
+export const PATCH = async (
+  req: NextRequest,
+  {
+    params,
+  }: {
+    params: Promise<{ id: string }>
+  },
+) => {
+  const { id } = await params
+  const semesterService = new SemesterService()
+  try {
+    const parsedBody = UpdateSemesterRequestBody.parse(await req.json())
+    const semester = await semesterService.updateSemester(id, parsedBody)
+    return NextResponse.json({ data: semester })
+  } catch (error) {
+    if (error instanceof ZodError) {
+      return NextResponse.json(
+        { error: 'Invalid request body', details: error.flatten() },
+        { status: StatusCodes.BAD_REQUEST },
+      )
+    } else if (error instanceof NotFound) {
+      return NextResponse.json({ error: 'Semester not found' }, { status: StatusCodes.NOT_FOUND })
+    } else {
+      console.error(error)
+      return NextResponse.json(
+        { error: 'Internal server error' },
+        { status: StatusCodes.INTERNAL_SERVER_ERROR },
+      )
+    }
+  }
+}
 
 /**
  * DELETE method to delete a semester.
