+++ src/app/api/projects/[id]/route.ts
@@ -1,0 +1,31 @@
+import { NextRequest, NextResponse } from 'next/server'
+import { StatusCodes } from 'http-status-codes'
+import ProjectService from '@/data-layer/services/ProjectService'
+import { NotFound } from 'payload'
+
+/**
+ * Fetches a project by its ID.
+ * @param req - The request object.
+ * @param params - The parameters object containing the project ID.
+ */
+
+export const GET = async (
+  req: NextRequest,
+  { params }: { params: Promise<{ id: string }> },
+): Promise<Response> => {
+  const { id } = await params
+  const projectService = new ProjectService()
+
+  try {
+    const data = await projectService.getProjectById(id)
+    return NextResponse.json({ data })
+  } catch (error) {
+    if (error instanceof NotFound) {
+      return NextResponse.json({ error: 'Project not found' }, { status: StatusCodes.NOT_FOUND })
+    }
+    return NextResponse.json(
+      { error: 'Internal Server Error' },
+      { status: StatusCodes.INTERNAL_SERVER_ERROR },
+    )
+  }
+}
