+++ src/app/api/projects/route.test.ts
@@ -7,7 +7,7 @@ import {
 } from '@/test-config/utils'
 import ProjectService from '@/data-layer/services/ProjectService'
 import { projectCreateMock, projectCreateMock2 } from '@/test-config/mocks/Project.mock'
-import { GET, POST } from '@/app/api/projects/route'
+import { GET, POST } from './route'
 
 describe('test /api/projects', () => {
   afterEach(async () => {
@@ -60,7 +60,7 @@ describe('test /api/projects', () => {
   })
 
   it('should create a project', async () => {
-    const req = createMockNextPostRequest('https://localhost:3000/api/projects', projectCreateMock)
+    const req = createMockNextPostRequest('', projectCreateMock)
     const res = await POST(req)
     expect(res.status).toBe(StatusCodes.CREATED)
     const project = (await res.json()).data
@@ -78,7 +78,7 @@ describe('test /api/projects', () => {
   })
 
   it('should fail to create a project', async () => {
-    const req = createMockNextPostRequest('https://localhost:3000/api/projects', {
+    const req = createMockNextPostRequest('', {
       ...projectCreateMock,
       description: undefined,
     })
