+++ src/app/api/projects/route.ts
@@ -31,7 +31,6 @@ export const GET = async (req: NextRequest): Promise<Response> => {
  * @param req - The request object.
  * @returns A JSON response containing the created project.
  */
-
 export const POST = async (req: NextRequest): Promise<Response> => {
   const projectService = new ProjectService()
