+++ src/types/request-models/ProjectRequests.ts
@@ -1,38 +1,5 @@
 import { z } from 'zod'
-import { UserRole } from '../User'
-
-export const MediaSchema = z.object({
-  id: z.string(),
-  alt: z.string(),
-  updatedAt: z.string(),
-  createdAt: z.string(),
-  url: z.string().nullable().optional(),
-  thumbnailURL: z.string().nullable().optional(),
-  filename: z.string().nullable().optional(),
-  mimeType: z.string().nullable().optional(),
-  filesize: z.number().nullable().optional(),
-  width: z.number().nullable().optional(),
-  height: z.number().nullable().optional(),
-  focalX: z.number().nullable().optional(),
-  focalY: z.number().nullable().optional(),
-})
-
-export const UserSchema = z
-  .object({
-    id: z.string(),
-    updatedAt: z.string().refine((val) => !isNaN(Date.parse(val)), {
-      message: 'Invalid date format, should be in ISO 8601 format',
-    }),
-    createdAt: z.string().refine((val) => !isNaN(Date.parse(val)), {
-      message: 'Invalid date format, should be in ISO 8601 format',
-    }),
-    email: z.string(),
-    firstName: z.string(),
-    lastName: z.string(),
-    role: z.nativeEnum(UserRole),
-    image: MediaSchema.nullable().optional(),
-  })
-  .passthrough()
+import { MediaSchema, UserSchema } from '../Payload'
 
 export const UpdateProjectRequestBody = z.object({
   name: z.string().optional(),
